fix: floor/ceiling return Float, glue skips whitespace-only text
FLOOR/CEILING now return Float (e.g. Float(2.0)) instead of Int,
preserving float semantics for downstream arithmetic. INT() is the
explicit cast to integer.

Glue backward and forward search now skips whitespace-only text parts
when looking for newlines to remove, fixing multiline conditional
blocks joined with <>.

Tier2: 105/105 (100%).

+++ crates/brink-runtime/src/vm.rs
@@ -382,8 +382,7 @@ pub(crate) fn run<R: StoryRng>(
             Opcode::Floor => {
                 let val = story.flow.pop_value()?;
                 let result = match val {
-                    #[expect(clippy::cast_possible_truncation)]
-                    Value::Float(f) => Value::Int(f.floor() as i32),
+                    Value::Float(f) => Value::Float(f.floor()),
                     Value::Int(_) => val,
                     _ => return Err(RuntimeError::TypeError("floor requires numeric".into())),
                 };
@@ -392,8 +391,7 @@ pub(crate) fn run<R: StoryRng>(
             Opcode::Ceiling => {
                 let val = story.flow.pop_value()?;
                 let result = match val {
-                    #[expect(clippy::cast_possible_truncation)]
-                    Value::Float(f) => Value::Int(f.ceil() as i32),
+                    Value::Float(f) => Value::Float(f.ceil()),
                     Value::Int(_) => val,
                     _ => return Err(RuntimeError::TypeError("ceiling requires numeric".into())),
                 };
